Moved different UI element to their own files.
