+++ .gitignore
@@ -1,2 +1,3 @@
 env
 docs/_build
+*__pycache__*
