+++ spade/ui/editorwidget.py
@@ -1,0 +1,5 @@
+from PyQt5.QtWidgets import QTextEdit
+
+class SpadeEditorWidget(QTextEdit): # TODO: should be just QWidget
+    def __init__(self):
+        super().__init__()
